requirements.txt
appdirs==1.4.4
#asgiref==3.2.10
distlib==0.3.1
Django==3.1
filelock==3.0.12
gunicorn==20.0.4
mysqlclient==1.4.6
Pillow==8.0.0
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
uWSGI==2.0.19.1
virtualenv==20.0.30
requests==2.24.0
channels==2.4.0
twilio==6.48.0
textmagic==2.0.3
sinchsms==1.0.4